# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Staging Superapp API",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server (Docker)"
        }
    ],
    "paths": {
...
        "/api/v1/inventory/adjustments": {
            "post": {
                "tags": [
                    "Inventory Transactions"
                ],
                "summary": "Adjust inventory stock",
                "operationId": "8bf514ff48a72821e86606e5f3a80fd8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdjustStockRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock adjustment berhasil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Stock adjustment berhasil."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/transfers": {
            "get": {
                "tags": [
                    "Inventory Transactions"
                ],
                "summary": "Get all transfers list",
                "operationId": "b7b96d8feae850725a6412d7af942735",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar semua dokumen transfer."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inventory Transactions"
                ],
                "summary": "Transfer inventory stock (Transfer Out)",
                "operationId": "b592727d9882b71f5118f9b8d8073289",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransferStockRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer Out berhasil dibuat.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transfer Out berhasil dibuat, barang sedang dalam perjalanan (Transit)."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/transfers/{id}/in": {
            "post": {
                "tags": [
                    "Inventory Transactions"
                ],
                "summary": "Transfer In inventory stock",
                "operationId": "a08c1889bf955a54a17a1e78b1181860",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the transfer",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "received_by": {
                                        "type": "string",
                                        "example": "admin"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfer In berhasil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transfer In berhasil, stok telah masuk ke gudang tujuan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/transfers/transit": {
            "get": {
                "tags": [
                    "Inventory Transactions"
                ],
                "summary": "Get list of transit transfers",
                "operationId": "91dc69baed981a69c27ae4a714852e49",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar barang dalam perjalanan (Transit)."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/transfers/{id}": {
            "get": {
                "tags": [
                    "Inventory Transactions"
                ],
                "summary": "Get transfer details",
                "operationId": "f50276dfe6660ec4261cd56fc9d1d69c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail transfer."
                    },
                    "404": {
                        "description": "Transfer tidak ditemukan."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/inventory/putaway": {
            "post": {
                "tags": [
                    "Inventory Transactions"
                ],
                "summary": "Putaway inventory stock",
                "operationId": "74ddf325dfb24ab3866eadef70e9c2bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PutawayStockRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Putaway berhasil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Putaway berhasil."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdjustStockRequest": {
                "required": [
                    "item_id",
                    "location_id",
                    "qty",
                    "created_by"
                ],
                "properties": {
                    "item_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "location_id": {
                        "type": "string",
                        "example": 1
                    },
                    "bin_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "nullable": true
                    },
                    "batch_no": {
                        "type": "string",
                        "example": "BATCH-001",
                        "nullable": true
                    },
                    "serial_no": {
                        "type": "string",
                        "example": "SN-001",
                        "nullable": true
                    },
                    "expired_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-12-31",
                        "nullable": true
                    },
                    "qty": {
                        "type": "integer",
                        "example": 10
                    },
                    "created_by": {
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            },
            "PutawayStockRequest": {
                "required": [
                    "item_id",
                    "location_id",
                    "source_bin_id",
                    "destination_bin_id",
                    "qty",
                    "created_by"
                ],
                "properties": {
                    "item_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "location_id": {
                        "type": "string",
                        "example": 1
                    },
                    "source_bin_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "destination_bin_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "660e8400-e29b-41d4-a716-446655440001"
                    },
                    "batch_no": {
                        "type": "string",
                        "example": "BATCH-001",
                        "nullable": true
                    },
                    "serial_no": {
                        "type": "string",
                        "example": "SN-001",
                        "nullable": true
                    },
                    "expired_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-12-31",
                        "nullable": true
                    },
                    "qty": {
                        "type": "integer",
                        "example": 50
                    },
                    "created_by": {
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            },
            "TransferStockRequest": {
                "required": [
                    "item_id",
                    "source_location_id",
                    "destination_location_id",
                    "qty",
                    "created_by"
                ],
                "properties": {
                    "item_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "source_location_id": {
                        "type": "string",
                        "example": 1
                    },
                    "source_bin_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000",
                        "nullable": true
                    },
                    "destination_location_id": {
                        "type": "string",
                        "example": 2
                    },
                    "destination_bin_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "660e8400-e29b-41d4-a716-446655440001",
                        "nullable": true
                    },
                    "batch_no": {
                        "type": "string",
                        "example": "BATCH-001",
                        "nullable": true
                    },
                    "serial_no": {
                        "type": "string",
                        "example": "SN-001",
                        "nullable": true
                    },
                    "expired_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-12-31",
                        "nullable": true
                    },
                    "qty": {
                        "type": "integer",
                        "example": 20
                    },
                    "created_by": {
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}